# docs/handoff/01_postanovka/04_openapi.yaml
openapi: 3.1.0
info:
  title: reHome Knowledge Base API
  version: 1.0.0
  summary: API модуля базы знаний платформы reHome
  description: "Публичный API модуля базы знаний reHome, через который платформа rehome.one\nи другие потребители получают\
    \ доступ к статьям help-центра, документам,\nAI-чату, поиску и карточкам квартир.\n\nПолное описание архитектуры и принципов\
    \ — в документе\n«API базы знаний reHome — ТЗ v1.0».\n\n## Аутентификация\nПоддерживаются два режима:\n- **m2m** (server-to-server):\
    \ OAuth 2.0 Client Credentials через kb-auth Keycloak,\n  Authorization: Bearer <jwt>\n- **user** (browser-to-server):\
    \ HttpOnly Cookie kb_session с JWT,\n  устанавливается при логине пользователя в rehome.one\n\n## Контроль доступа\nКаждый\
    \ ресурс имеет access_level (PUBLIC / LOGGED / AGENT / STAFF / LEGAL / HR_RESTRICTED).\nscope пользователя вычисляется\
    \ бэкендом из проверенного токена,\nНИКОГДА не передаётся клиентом. Фильтрация применяется на уровне хранилища.\n\n##\
    \ Версионирование\nВерсия в URL: /api/v1/. Breaking changes требуют новой мажорной версии\nи параллельной работы старой\
    \ минимум 6 месяцев.\n"
  contact:
    name: reHome Engineering
    email: engineering@rehome.one
    url: https://rehome.one
  license:
    name: Proprietary
  termsOfService: https://rehome.one/legal/api-terms
servers:
- url: https://api.rehome.one/kb
  description: Production
- url: https://api.staging.rehome.one/kb
  description: Staging
...
security:
- BearerAuth: []
- CookieAuth: []
...
paths:
...
  /api/v1/collaborators:
    get:
      tags:
      - Collaborators
      summary: Список коллаборантов
      description: |
        Возвращает список внешних коллаборантов (УК/ТСЖ, аварийные службы, клининг,
        переезды, ремонт, страховые и т.д.). Подробно — в ПЗ «База знаний v1.4»
        раздел 10.

        Фильтрация по типу, статусу, географии. Видимость зависит от scope
        (ADR-0014 §3):
        - guest / tenant / landlord / agent: только D-группа (management_company,
          emergency_service) — публичные контакты
        - staff_*: все группы (A/B/C/D)

        **Slice 1 (#231-232 landed):** реализованы фильтры по `type`, `status`,
        `service_area` (ILIKE substring). Фильтры `financial_group`, `city`,
        `district`, `premises_id` — Slice 5 (PremisesCollaborator junction).
      operationId: listCollaborators
      x-implementation-status: implemented
      x-implementation-slice: 1
      x-adr: ADR-0014
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: type
        in: query
        schema:
          $ref: '#/components/schemas/CollaboratorType'
      - name: financial_group
        in: query
        schema:
          type: string
          enum:
          - A
          - B
          - C
          - D
      - name: status
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - PENDING_REVIEW
          - ACTIVE
          - SUSPENDED
          - ARCHIVED
      - name: city
        in: query
        schema:
          type: string
      - name: district
        in: query
        schema:
          type: string
      - name: premises_id
        in: query
        description: Вернуть только коллаборантов, обслуживающих этот объект
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/CollaboratorSummary'
                    pagination:
                      $ref: '#/components/schemas/Pagination'
    post:
      tags:
      - Collaborators
      summary: Создать коллаборанта
      description: |
        Создаёт коллаборанта в статусе DRAFT по умолчанию. D-группа
        (management_company, emergency_service) — auto-ACTIVE (ТЗ §3.10.1).
        `financial_group` auto-derive из `type` (кроме `type='other'` —
        обязательно указывать explicitly). Инвариант (type, financial_group)
        per ТЗ §10.3 enforced через CHECK constraint в БД (ADR-0014 §2).

        Требует scope = staff_support+.
      operationId: createCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 1
      x-adr: ADR-0014
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaboratorInput'
      responses:
        '201':
          description: Создано (статус — DRAFT или PENDING_REVIEW)
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Collaborator'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/collaborators/{collaborator_id}:
    parameters:
    - name: collaborator_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Collaborators
      summary: Карточка коллаборанта
      description: |
        Состав возвращаемых полей зависит от scope (ADR-0014 §7):
        - guest / tenant / landlord / agent: CollaboratorPublic — type, brand_name,
          financial_group, status, service_area, working_hours, website, rating
          (БЕЗ юр.реквизитов / contacts.person_name / financial_terms)
        - staff_*: CollaboratorInternal — + name, inn/ogrn/kpp, contacts (полные),
          financial_terms, api_integration, sla, counterparty_check
        - staff_admin (STAFF + LEGAL): + audit_log

        404 при out-of-scope (anti-enumeration mask, ADR-0003).
      operationId: getCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 1
      x-adr: ADR-0014
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Collaborator'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Collaborators
      summary: Обновить карточку коллаборанта
      description: |
        Partial update. `type` и `financial_group` могут меняться, но pair
        invariant (ТЗ §10.3, ADR-0014 §2) валидируется в Pydantic — несовпадение
        → 422. Status можно менять в любую сторону (transition validation отложен
        до Slice 2: `/activate`, `/suspend`).

        Требует STAFF+ scope.
      operationId: patchCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 1
      x-adr: ADR-0014
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaboratorInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Collaborator'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - Collaborators
      summary: Архивировать коллаборанта (soft delete)
      description: |
        Soft delete — status → ARCHIVED. Привязки к объектам сохраняются
        (Slice 5 PremisesCollaborator). Audit row пишется с
        `metadata.previous_status`.

        Требует STAFF+ scope.
      operationId: archiveCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 1
      x-adr: ADR-0014
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Архивировано
  /api/v1/collaborators/{collaborator_id}/activate:
    post:
      tags:
      - Collaborators
      summary: Активация коллаборанта (DRAFT/PENDING_REVIEW → ACTIVE)
      description: 'Активация возможна только когда: counterparty_check = CLEAN,

        contract_document_id заполнен (для групп A/B/C), responsible_internal назначен.

        Любая активация фиксируется в аудит-логе.

        '
      operationId: activateCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 2
      x-adr: ADR-0014
      parameters:
      - name: collaborator_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Активировано
        '422':
          description: Не выполнены условия активации
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/collaborators/{collaborator_id}/suspend:
    post:
      tags:
      - Collaborators
      summary: Приостановка коллаборанта
      description: 'Временная остановка работы с коллаборантом без архивации.

        Например, при просрочке проверки контрагента или сбое API партнёра.

        '
      operationId: suspendCollaborator
      x-implementation-status: implemented
      x-implementation-slice: 2
      x-adr: ADR-0014
      parameters:
      - name: collaborator_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - reason
              properties:
                reason:
                  type: string
                until:
                  type: [string, 'null']
                  format: date
      responses:
        '200':
          description: Приостановлено
...
  /api/v1/premises/{premises_id}/collaborators:
    parameters:
    - name: premises_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Collaborators
      summary: Коллаборанты, обслуживающие данный объект
      description: 'Возвращает список PremisesCollaborator-связей: УК, аварийная служба,

        рекомендуемые сантехники, привязанные к этому объекту.

        Видимость публичных контактов — для tenant данной квартиры.

        '
      operationId: listPremisesCollaborators
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/PremisesCollaborator'
    post:
      tags:
      - Collaborators
      summary: Назначить коллаборанта на объект
      description: Требует staff_support+
      operationId: assignCollaboratorToPremises
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PremisesCollaboratorInput'
      responses:
        '201':
          description: Назначено
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PremisesCollaborator'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Bearer JWT — m2m или user токен.

        m2m: получить через kb-auth Client Credentials Grant.

        '
    CookieAuth:
      type: apiKey
      in: cookie
      name: kb_session
      description: HttpOnly Secure cookie, ставится при логине в rehome.one
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      schema:
        type: string
        format: uuid
      description: UUID запроса для трассировки. Если не передан — генерируется.
...